Simplify Vendor interface to keep a single BuylistEntry
We're about to switch to the official ABUGames API, which only has
a single condition, like every other vendor, so it becomes less
interesting to keep this notation around.

This change makes the overall code a bit simpler, like dropping the
need to carry over the MINT condition or checking against multiple
conditions in arbitrage.