tcgplayer: Add retry calls for an odd API glitch
API responds with 200 but returns an html page containing a 403 error.
It is speculated that the auth token is expiring during the run,
something is noticing, and fixing it by refreshing it or obtaining a
fresh one, failing the current call.
So instead just retry the call for 5 times if that happens.