mtgmatcher: Parse timeshifted editions in a different way
Since some provides lump together all the timeshifted cards, check for
that keyword presence, and assign the correct edition with MatchInSet().
This allows us to drop the old TSB special cases, but we still need to
add an entry for cardmarket odd edition split (Extras would conflict with
later parsing for CBB sets.