mtgmatcher/filter: Support both showcase and borderless on the same card
It was becoming more and more common that showcase and borderless
tags were applied on the same card object, but the input card only
carried one of them. Having them in the same filter loop was causing
problems due to being too aggressive, while now it is only performed
when filtering was not aggressive enough.