sealedev: Skip errors coming from Sample Pack
These packs are usually filled in last, but they are found in
several other products (namely commander decks) - an error in
the calc for a missing pack would result in the whole ev being
discarded. Instead, given the fact that Sample Packs don't affect
ev much, ignore them in case they error, and show as much data
as possible for container products.