mtgmatcher/api: Access the names map directly in SearchEquals()
Now that all portions of the name are correctly loaded in, we don't need
to pass a slice and perform an o(N) operation, but for the most common
case we can optimize as an o(1).

Keep SearchSealedEquals separate in case there are search operations to
be performed only sealed products, but make is similar to the singles
one by returning all available hashes if query is empty.

(before) BenchmarkSearchEquals-14        349068      3460.0 ns/op
(after)  BenchmarkSearchEquals-14       8669376       140.3 ns/op