mtgban: Ignore data presence when unfolding
It's important to expose the fact that a scraper contains no data, but
the UnfoldScrapers() shadows this fact and skips the subscraper entirely.
Whether a retail or buylist section of a scraper has been run is already
identified by the timestamp, therefore we can preserve the lack of data
as is.